+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 #### 3.5.12
 - fix: node:click is triggered twice while clicking a node;
-- fix: update combo edge when drag node out of it problem.
+- fix: update combo edge when drag node out of it problem;
+- feat: animate configuration for combo, true by default.
 
 #### 3.5.11
 - feat: graph.priorityState api;